Material: material main module, move card related defs into mat utils module
